requirements.txt
google-generativeai==0.8.3
gradio==5.12.0
python-dotenv==1.0.1
